+++ example/a.ts
@@ -1,21 +1,21 @@
 import { Elysia, t } from '../src'
+import { req } from '../test/utils'
 
-const app = new Elysia()
-	.derive(({ cookie: { test } }) => {
-		if (!test.value) {
-			test.value = 'Hello, world!'
-		}
-
-		return {}
+const api = new Elysia().get('/', ({ query }) => query, {
+	query: t.Object({
+		date: t.Date()
 	})
-	.get('/', () => 'Hello, world!')
+})
 
-app.handle(
-	new Request('http://localhost:3000/', {
-		headers: {
-			cookie: 'test=Hello, world!'
-		}
-	})
-)
-	.then((x) => x.headers)
-	.then(console.log)
+api.handle(req(`/?date=${Date.now()}`)).then(x => x.json()).then(console.log)
+
+// const app = new Elysia()
+// 	.get('/', () => 'ok', {
+// 		query: t.Object({
+// 			key1: t.Union([t.Array(t.String()), t.String()])
+// 		})
+// 	})
+
+// app.handle(req('/?key1=ab&key1=cd&z=が'))
+// 	.then((x) => x.status)
+// 	.then(console.log)