+++ src/type-system.ts
@@ -369,12 +369,22 @@ export const ElysiaType = {
 						t.String({
 							format: 'date-time',
 							default: new Date().toISOString()
-						})
+						}),
+						t.Number()
 					],
 					property
 				)
 			)
 			.Decode((value) => {
+				if(typeof value === "number") {
+					const date = new Date(value)
+
+					if (!Value.Check(schema, date))
+						throw new ValidationError('property', schema, date)
+
+					return date
+				}
+
 				if (value instanceof Date) return value
 
 				const date = new Date(value)